pyproject.toml
[tool.poetry]
name = "sharecipe"
version = "0.1.0"
description = "sharecipe"
authors = ["JunHCha <quffl9612@naver.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3"
djangorestframework = "^3.12.4"
drf-spectacular = "^0.21.0"
psycopg = "^3.0.5"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
